requirements.txt
requests 
tqdm
Pillow
geopy
pandas
msgpack
kaggle
huggingface_hub
flickr-api
python-dotenv
geopandas
shapely
hdbscan
scikit-learn
torch
torchvision
timm
streetview
requests
datasets
modal
sparqlwrapper